fix: type parameter deleteById()

+++ backend/src/main/java/org/tarjetamish/contact/repository/impl/ContactRepository.java
@@ -17,5 +17,5 @@ public interface ContactRepository {
 
     int save(Contact contact);
 
-    int deleteById(Long id);
+    int deleteById(int id);
 }
